Save students and Generic Way to get fields
* The student modifications can now be saved. The solution is to
split the treatment for the personal and the medical information.
Code duplication has been done with generic methods.

* No more need to ask for a specific list of properties when
displaying either a student or a teacher. If no property list is
empty or null, the server will read the ontology to get the list
of all properties defined for this type.

* Add code coverage of the unit-tests. The whole OntologyManager
is now covered.