Business Component, Unit-Tests and Classes/Years
* A level has been added into the design: the business component. The
service level handles now only the transaction. The business component
level handles the logic of the functionality when the manager handle a
single action. The BC logic uses a combination of the actions provided by
the manager level.

* The unit-test system has been improved to cover most of the code and
branches from the services to the managers.

* The base functionalities of the year and class have been implemented. A
year can be started and stopped. Classes can be created and associated
with a year. The gui has been updated to enable these features and prevent
some actions such as create a class when a year has not been started. This
limitation has not been implemented into the webservices yet but will be.
The years and classes' list is now displayed (but not ordered).